Add descriptions to each figure
