+++ pf2/figures/commonFuncs/plotGeneral.py
@@ -40,3 +40,15 @@ def plot_avegene_per_status(
     ax.set(ylabel=f"Average {gene}")
 
     return df
+
+
+def rotate_xaxis(ax, rotation=90):
+    """Rotates text by 90 degrees for x-axis"""
+    ax.set_xticks(ax.get_xticks())
+    ax.set_xticklabels(labels=ax.get_xticklabels(), rotation=rotation)
+
+
+def rotate_yaxis(ax, rotation=90):
+    """Rotates text by 90 degrees for y-axis"""
+    ax.set_yticks(ax.get_yticks())
+    ax.set_yticklabels(labels=ax.get_yticklabels(), rotation=rotation)
